pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "identity-access-service"
version = "0.1.0"
description = "Production-style IAM backend with FastAPI, PostgreSQL, JWT auth, RBAC, and multi-tenancy."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.13,<2.0",
    "fastapi>=0.115,<1.0",
    "psycopg[binary]>=3.2,<4.0",
    "pydantic-settings>=2.6,<3.0",
    "sqlalchemy>=2.0,<3.0",
    "uvicorn[standard]>=0.30,<1.0",
]

[project.optional-dependencies]
auth = [
    "pyjwt>=2.10,<3.0",
    "pwdlib[argon2]>=0.2,<1.0",
]
dev = [
    "httpx>=0.28,<1.0",
    "pytest>=8.3,<9.0",
    "ruff>=0.8,<1.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
addopts = "-q -p no:cacheprovider -p no:asyncio"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["B", "E", "F", "I", "N", "UP"]
ignore = ["B008"]
